requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2

# AI/ML - FREE LOCAL MODELS
transformers==4.45.0
torch==2.1.1
accelerate==0.34.0
peft==0.12.0  # For LoRA fine-tuning
bitsandbytes==0.44.0  # For 4-bit/8-bit quantization
sentencepiece==0.2.0  # For tokenization
sacremoses==0.1.1  # For multilingual support

# Image Generation - FREE LOCAL
diffusers==0.30.0  # Stable Diffusion
html2image==2.0.4.3  # Fallback for text posters
Pillow==10.1.0

# Training & Evaluation
datasets==2.19.0
evaluate==0.4.1
rouge-score==0.1.2
sacrebleu==2.4.0
bert-score==0.3.13  # For BERTScore evaluation
nltk==3.8.1  # For BLEU score
pandas==2.1.3  # For dataset handling

# Gemini API (for content polishing)
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
numpy==1.24.3
